+++ src/routes/upload.routes.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import deserializeUser from '../middleware/deserializeUser';
 import requireUser from '../middleware/requireUser';
 import { uploadController } from '../controllers/upload.controller';
-import upload from '../config/storageConfig';
+import upload from '../middleware/multarMiddleware';
